Setup backwards compatible settings with jest-preset
